# Optimized token chunking with PyTorch's unfold instead of loop slicing.